requirements.txt
art
colorama
discord.py-self==1.7.3
Faker
pypresence
requests
pillow
pystyle
pyinstaller==5.11.0